[Banners] Extract to separate component
