+++ hosting/src/global/Home/Home.html
@@ -1,34 +1,6 @@
 <!-- Banner -->
 <div class="bg-gray-400 w-full">
-  <div
-    v-if="banners.length > 0"
-    class="relative max-w-7xl m-auto overflow-hidden relative"
-  >
-    <div class="aspect-w-16 aspect-h-9 w-full overflow-hidden">
-      <div
-        v-for="(banner, i) in banners"
-        :key="'slide' + i"
-        @click="bannerClick(banner)"
-        class="absolute bg-cover bg-top inset-0 w-full h-full transition-all ease-in-out duration-1000 transform slide object-contain"
-        :class="{ ['translate-x-0']: i === 0, ['translate-x-full']: i > 0 }"
-        :style="{ backgroundImage: 'url('+banner.banner169Url+')' }"
-      ></div>
-    </div>
-    <div
-      v-if="banners.length > 1"
-      class="absolute bottom-0 left-0 rounded-tr-md sm:rounded-tr-xl lg:rounded-tr-3xl bg-black bg-opacity-30 pr-1 w-8 h-8 sm:w-10 sm:h-10 md:w-12 md:h-12 lg:w-14 lg:h-14 text-lg sm:text-xl md:text-2xl lg:text-4xl flex items-center justify-center text-white cursor-pointer"
-      @click="previousSlide(); resetSlideTimer()"
-    >
-      &#x276E;
-    </div>
-    <div
-      v-if="banners.length > 1"
-      class="absolute bottom-0 right-0 rounded-tl-md sm:rounded-tl-xl lg:rounded-tl-3xl bg-black bg-opacity-30 pl-1 w-8 h-8 sm:w-10 sm:h-10 md:w-12 md:h-12 lg:w-14 lg:h-14 text-lg sm:text-xl md:text-2xl lg:text-4xl flex items-center justify-center text-white cursor-pointer"
-      @click="nextSlide(); resetSlideTimer()"
-    >
-      &#x276F;
-    </div>
-  </div>
+  <HomeBanners />
 </div>
 
 <!-- Calendar lookahead -->
