+++ hosting/src/global/Home/HomeBanners.vue
@@ -1,0 +1,2 @@
+<template src="./HomeBanners.html"></template>
+<script src="./HomeBanners.ts" lang="ts"></script>
